video: mxc_hdmi/mxc_edid: Introduce fractional modes
In order to support fractional refresh rates (23.976, 29.97, 59.94, 119.88Hz)
we create 'synthetic' modes. These have the same timing as the corresponding
non-fractional modes (i.e. 24, 30, 60 and 120Hz). However, the refresh rate
is set to 23, 29, 59 and 119, respectively. In addition they have the mode
bit FB_VMODE_FRACTIONAL set. This bit is taken in acount when calculating
the real pixel clock.

Signed-off-by: Rudi <r.ihle@s-t.de>